+++ pythx/__init__.py
@@ -4,8 +4,7 @@ __author__ = """Dominik Muhs"""
 __email__ = "dominik.muhs@consensys.net"
 __version__ = "0.1.0"
 
-
-from pythx.config import config
+from pythx.conf import config
 from pythx.api.client import Client
 from pythx.models.exceptions import (
     PythXBaseException,
