+++ pythx/api/handler.py
@@ -6,8 +6,6 @@ import requests
 
 from pythx import config
 from pythx.middleware.base import BaseMiddleware
-from pythx.models import request as reqmodels
-from pythx.models import response as respmodels
 from pythx.models.exceptions import PythXAPIError
