+++ pythx/conf/__init__.py
@@ -1,0 +1,4 @@
+from pythx.conf.base import PythXConfig
+
+
+config = PythXConfig()
