--- pythx/config/__init__.py
@@ -1,4 +1,0 @@
-from pythx.config.base import PythXConfig
-
-
-config = PythXConfig()
