+++ pythx/models/request/auth_logout.py
@@ -1,12 +1,7 @@
-import json
-from datetime import datetime
-from typing import Any, Dict, List
+from typing import Dict
 
-import dateutil.parser
-
-from pythx.models.exceptions import RequestDecodeError, RequestValidationError
+from pythx.models.exceptions import RequestDecodeError
 from pythx.models.request.base import BaseRequest
-from pythx.models.util import dict_delete_none_fields
 
 
 class AuthLogoutRequest(BaseRequest):
