+++ pythx/models/request/version.py
@@ -1,13 +1,4 @@
-import json
-from datetime import datetime
-from typing import Any, Dict, List
-
-import dateutil.parser
-
-from pythx.models.exceptions import RequestDecodeError, RequestValidationError
-from pythx.models.request.analysis_status import AnalysisStatusRequest
 from pythx.models.request.base import BaseRequest
-from pythx.models.util import dict_delete_none_fields
 
 
 class VersionRequest(BaseRequest):
