+++ pythx/models/response/analysis_submission.py
@@ -1,10 +1,5 @@
-import json
-from typing import Any, Dict, List
-
-from pythx.models.exceptions import ResponseDecodeError
 from pythx.models.response.analysis import Analysis
 from pythx.models.response.base import BaseResponse
-from pythx.models.response.issue import Issue, SourceFormat, SourceType
 
 
 class AnalysisSubmissionResponse(BaseResponse):
