+++ pythx/models/response/detected_issues.py
@@ -1,8 +1,6 @@
-import json
 from typing import Any, Dict, List
 
 from pythx.models.exceptions import ResponseDecodeError
-from pythx.models.response.analysis import Analysis
 from pythx.models.response.base import BaseResponse
 from pythx.models.response.issue import Issue, SourceFormat, SourceType
