+++ tests/test_config.py
@@ -1,7 +1,7 @@
 import pytest
 
 from pythx import config
-from pythx.config.base import PythXConfig
+from pythx.conf.base import PythXConfig
 
 TEST_KEY = "test_key"
 TEST_VALUE = ("test_method", "test_value")
